+++ kolibri/plugins/coach/assets/src/views/groups-page/create-group-modal.vue
@@ -94,7 +94,7 @@
         return '';
       },
       nameIsInvalid() {
-        return !!this.nameIsInvalidText;
+        return Boolean(this.nameIsInvalidText);
       },
       formIsValid() {
         return !this.nameIsInvalid;