+++ kolibri/plugins/coach/assets/src/views/lessons/LessonsRootPage.vue
@@ -1,7 +1,7 @@
 <template>
 
   <div>
-    <h1>{{ $tr('classLessons', { className }) }}</h1>
+    <h1>{{ $tr('classLessons') }}</h1>
     <div class="filter-and-button">
       <k-select
         :label="$tr('show')"
@@ -42,14 +42,14 @@
           <td class="core-table-icon-col">
             <content-icon :kind="lessonKind" />
           </td>
-          <td class="core-table-main-col">
+          <td class="core-table-main-col lesson-title-col">
             <k-router-link
               :to="lessonSummaryLink({ lessonId: lesson.id, classId })"
-              :text="lesson.name"
+              :text="lesson.title"
             />
           </td>
           <td>{{ $tr('numberOfResources', { count: lesson.resources.length }) }}</td>
-          <td>{{ getLessonVisibility(lesson.assigned_groups) }}</td>
+          <td>{{ getLessonVisibility(lesson.lesson_assignments) }}</td>
           <td>
             <status-icon :active="lesson.is_active" />
           </td>
@@ -157,7 +157,7 @@
       },
     },
     $trs: {
-      classLessons: '{ className } lessons',
+      classLessons: 'Lessons',
       show: 'Show',
       allLessons: 'All lessons',
       activeLessons: 'Active lessons',
@@ -172,11 +172,11 @@
       status: 'Status',
       numberOfResources:
         '{count, number, integer} {count, plural, one {resource} other {resources}}',
-      noLessons: 'No lessons',
+      noLessons: 'You do not have any lessons',
       noActiveLessons: 'No active lessons',
       noInactiveLessons: 'No inactive lessons',
       lessonStatusDescription: 'Lesson status description',
-      statusTooltipText: 'Active: learners can see lesson. Inactive: hidden from learners.',
+      statusTooltipText: 'Learners can only see active lessons',
     },
   };
 
@@ -184,6 +184,9 @@
 
 
 <style lang="stylus" scoped>
+
+  .lesson-title-col
+    width: 40%
 
   .filter-and-button
     display: flex