+++ kolibri/plugins/facility_management/assets/src/views/manage-class-page/class-create-modal.vue
@@ -96,7 +96,7 @@
         return '';
       },
       nameIsInvalid() {
-        return !!this.nameIsInvalidText;
+        return Boolean(this.nameIsInvalidText);
       },
       formIsValid() {
         return !this.nameIsInvalid;