+++ kolibri/plugins/learn/assets/src/views/points-popup/index.vue
@@ -11,7 +11,6 @@
 
 
     <div class="points-wrapper">
-      <!-- <h2>{{ $tr('pointsForCompletion') }}</h2> -->
       <div class="points">
         <points-icon class="points-icon" :active="true" />
         <span class="points-amount">{{ $tr('plusPoints', { maxPoints }) }}</span>
@@ -26,11 +25,11 @@
       {{ $tr('signIn') }}
     </ui-alert>
 
-    <div class="next-item-section">
+    <div v-if="nextContent" class="next-item-section">
       <h2 class="next-item-heading">{{ $tr('nextContent') }}</h2>
       <div>
-        <content-icon class="nex-item-icon" :kind="kind" />
-        <span class="next-item-title">{{ title }}</span>
+        <content-icon class="nex-item-icon" :kind="nextContent.kind" />
+        <span class="next-item-title">{{ nextContent.title }}</span>
       </div>
     </div>
 
@@ -58,6 +57,16 @@
   import kButton from 'kolibri.coreVue.components.kButton';
   import uiAlert from 'keen-ui/src/UiAlert';
 
+  const kindToLabelMap = {
+    [ContentNodeKinds.TOPIC]: 'topic',
+    [ContentNodeKinds.CHANNEL]: 'channel',
+    [ContentNodeKinds.EXERCISE]: 'exercise',
+    [ContentNodeKinds.VIDEO]: 'video',
+    [ContentNodeKinds.AUDIO]: 'audio',
+    [ContentNodeKinds.DOCUMENT]: 'document',
+    [ContentNodeKinds.HTML5]: 'html5',
+  };
+
   export default {
     name: 'pointsPopup',
     $trs: {
@@ -65,11 +74,12 @@
       niceWork: 'Great work! Keep it up!',
       nextContent: 'Next Item',
       topic: 'Topic',
+      channel: 'Channel',
       exercise: 'Exercise',
       video: 'Video',
       audio: 'Audio',
       document: 'Document',
-      html5: 'HTML5 app',
+      html5: 'App',
       item: 'Item',
       close: 'Close',
       pointsForCompletion: 'Points for completion',
@@ -90,29 +100,16 @@
       },
     },
     props: {
-      kind: { type: String },
-      title: { type: String },
+      nextContent: {
+        type: Object,
+      },
     },
     computed: {
       maxPoints() {
         return MaxPointsPerContent;
       },
       nextKind() {
-        const kind = this.kind;
-        if (kind === ContentNodeKinds.TOPIC) {
-          return this.$tr('topic');
-        } else if (kind === ContentNodeKinds.EXERCISE) {
-          return this.$tr('exercise');
-        } else if (kind === ContentNodeKinds.VIDEO) {
-          return this.$tr('video');
-        } else if (kind === ContentNodeKinds.AUDIO) {
-          return this.$tr('audio');
-        } else if (kind === ContentNodeKinds.DOCUMENT) {
-          return this.$tr('document');
-        } else if (kind === ContentNodeKinds.HTML5) {
-          return this.$tr('html5');
-        }
-        return this.$tr('item');
+        return this.$tr(kindToLabelMap[this.nextContent.kind] || 'item');
       },
     },
     methods: {