+++ kolibri/plugins/setup_wizard/assets/src/views/onboarding-forms/facility-permissions-form/index.vue
@@ -64,22 +64,14 @@
 
     <onboarding-form
       :header="$tr('facilityPermissionsSetupFormHeader')"
+      :description="$tr('facilityPermissionsSetupFormDescription')"
       :submitText="submitText"
       @submit="setPermissions"
     >
 
-      <template slot="description">
-        {{ $tr('facilityPermissionsSetupFormDescription') }}
-        <k-button
-          appearance="basic-link"
-          :text="$tr('facilityPermissionsPresetDetailsLink')"
-          ref="details"
-          @click="showFacilityPermissionsDetails"
-        />
-      </template>
-
       <label class="permission-preset">
         <k-radio-button
+          ref="first-button"
           class="permission-preset-radio-button"
           v-model="selectedPreset"
           radiovalue="nonformal"
@@ -114,6 +106,14 @@
         </span>
       </label>
 
+      <k-button
+        slot="footer"
+        appearance="basic-link"
+        :text="$tr('facilityPermissionsPresetDetailsLink')"
+        @click="showFacilityPermissionsDetails"
+      />
+
+
     </onboarding-form>
 
   </div>
@@ -141,8 +141,8 @@
     $trs: {
       facilityPermissionsSetupFormHeader: 'Choose a Facility setup',
       facilityPermissionsSetupFormDescription:
-        'How will you be using Kolibri? You can customize these settings later.',
-      facilityPermissionsPresetDetailsLink: 'Setup details',
+        'How will you be using Kolibri? (You can customize these settings later)',
+      facilityPermissionsPresetDetailsLink: 'More information about these settings',
       facilityPermissionsPresetDetailsHeader: 'Facility setup details',
       adminManagedSetupTitle: 'Admin-managed',
       adminManagedSetupDescription: 'For schools and other formal learning contexts',
@@ -175,7 +175,7 @@
       };
     },
     mounted() {
-      this.$refs.details.$el.focus();
+      this.$refs['first-button'].focus();
     },
     methods: {
       setPermissions() {