+++ kolibri/plugins/user/assets/src/state/mutations.js
@@ -5,6 +5,9 @@ export default {
   SET_PAGE_STATE(state, pageState) {
     state.pageState = pageState;
   },
+  SET_FACILITY_ID(state, facilityId) {
+    state.facilityId = facilityId;
+  },
   // Profile Page Mutations
   SET_PROFILE_BUSY(state, isBusy) {
     state.pageState.busy = isBusy;
@@ -16,6 +19,13 @@ export default {
     state.pageState.error = isError;
     state.pageState.errorMessage = errorMessage;
     state.pageState.errorCode = errorCode;
+  },
+  SET_PROFILE_PASSWORD_MODAL(state, isVisible) {
+    state.pageState.passwordState.modal = isVisible;
+  },
+  SET_PROFILE_PASSWORD_ERROR(state, { isError, errorMessage = '' }) {
+    state.pageState.passwordState.error = isError;
+    state.pageState.passwordState.errorMessage = errorMessage;
   },
   // Sign Up Page Mutations
   SET_SIGN_UP_BUSY(state, isBusy) {
@@ -32,6 +42,11 @@ export default {
       error: false,
       errorCode: null,
       errorMessage: '',
+      passwordState: {
+        modal: false,
+        success: false,
+        error: false,
+      },
     };
   },
   RESET_SIGN_UP_STATE(state) {